Update exercicio_1.py
allan km ex 1

+++ exercicio_1.py
@@ -19,7 +19,7 @@ Dica: Lembre do comando de resto da divisão inteira!
 """
 
 
-numero = float(input("Digite um número:"))
+numero = int(input("Digite um número:"))
 
 if numero % 2 == 0:
   print("Par")
